Home

Analytical estimation of $\alpha_{x}$ accounting for linkage. We explore the impact of linkage and background selection at positive selected alleles sites. The package solve anylitical approximations for different genetic scenarios in order to estimate the strenght and rate of adaptation.

When empircal values of polymorphim and divergence are given, they will be used to discern their expected correspoding values modeled under any Distribution of Fitness Effect ($DFE$) and background selection values ($B$).

Our goal is to subset summary statistics given a set of empirical values for any genetic scenario that would be used as prior distributions in $ABC$ algorithms.

Analytical.fixNeut — Function

fixNeut()

Expected neutral fixations rate reduce by a background selection value. It takes into accoun the ammount probability of being synonymous. Testing build

\[\mathbb{E}[D_{s}] = (1 - p_{-} - p_{+}) B \frac{1}{2N}\]

Returns

Expected rate of neutral fixations: Float64

Analytical.fixNegB — Function

fixNegB(ppos)

Expected fixation rate from negative DFE.

\[\mathbb{E}[D_{n-}] = p_{-}\left(2^-\alpha\beta^\alpha\left(-\zeta[\alpha,\frac{2+\beta}{2}] + \zeta[\alpha,1/2(2-\frac{1}{N+\beta})]\right)\right)\]

Arguments

ppos::Float64: selection coefficient

Returns

Expected rate of fixations from negative DFE: Float64

Analytical.pFix — Function

pFix()

Expected positive fixation rate

\[\mathbb{E}[D_{n+}] = p_{+} B (1 - 2^{(-2s)})\]

Arguments

ppos::Float64: selection coefficient

Returns

Rate of neutral fixation: Float64

Analytical.fixPosSim — Function

fixPosSim(gamma,ppos)

Expected positive fixations rate reduced due to the impact of background selection and linkage. The formulas presented at the approach have been subjected to several previous theoretical works (Charlesworth B., 1994, Hudson et al., 1995, (Nordborg et al. 1995)(https://doi.org/10.1017/S0016672300033619), Barton NH., 1995).

Then, the probabilty of fixation of positively selected alleles is reduced by a factor math \phi across all deleterious linked sites.

\[\phi(t,s) = \euler^{[\frac{-2\mu}{}]}\]

\[\Phi = \prod_{1}^{L} \phi(t,s)\]

\[\mathbb{E}[D_{n+}]' = \Phi \mathbb{E}[D_{n+}]\]

Arguments

ppos::Float64: selection coefficient

Returns

Expected rate of positive fixations under background selection: Float64

Analytical.DiscSFSNeutDown — Function

DiscSFSNeutDown()

Expected neutral allele frequency rate reduce by background selection. The spectrum depends on the number of individual: adap.NN.

\[\mathbb{E}[Ps_{(x)}] = \sum{x^{*}=x}{x^{*}=1}f_{B}(x)\]

Return:

- Array{Float64}(adap.NN · adap.B,1)