Analytical

Analytical approximation to $\alpha_{x}$ accounting for linkage. We explore the impact of linkage and background selection at positive selected alleles sites. The package solve anylitical approximations for different genetic scenarios in order to estimate the strenght and rate of adaptation.

When empircal values of polymorphim and divergence are given, they will be used to discern their expected correspoding values modeled under any Distribution of Fitness Effect (DFE) and background selection values (B).

Our goal is to subset summary statistics given a set of empirical values for any genetic scenario that would be used as prior distributions in ABC algorithms.

Analytical.fixNeut — Function

fixNeut()

Expected neutral fixations rate reduce by a background selection value.

\[\mathbb{E}[D_{s}] = (1 - p_{-} - p_{+}) B \frac{1}{2N}\]

Returns

Float64: expected rate of neutral fixations.

Analytical.fixNegB — Function

fixNegB(ppos)

Expected fixation rate from negative DFE.

\[\mathbb{E}[D_{n-}] = p_{-}\left(2^-\alpha\beta^\alpha\left(-\zeta[\alpha,\frac{2+\beta}{2}] + \zeta[\alpha,1/2(2-\frac{1}{N+\beta})]\right)\right)\]

Arguments

ppos::Float64: selection coefficient.

Returns

Float64: expected rate of fixations from negative DFE.

Analytical.pFix — Function

pFix()

Expected positive fixation rate.

\[\mathbb{E}[D_{n+}] = p_{+} \cdot B \cdot (1 - e^{(-2s)})\]

Arguments

ppos::Float64: selection coefficient.

Returns

Float64: expected rate of positive fixation.

Analytical.fixPosSim — Function

fixPosSim(gamma,ppos)

Expected positive fixations rate reduced due to the impact of background selection and linkage. The probabilty of fixation of positively selected alleles is reduced by a factor Φ across all deleterious linked sites Analytical.phiReduction.

\[\mathbb{E}[D_{n+}] = \Phi \cdot \mathbb{E}[D_{n+}]\]

Arguments

ppos::Float64: selection coefficient

Returns

Float64: expected rate of positive fixations under background selection

Analytical.DiscSFSNeutDown — Function

DiscSFSNeutDown()

Expected rate of neutral allele frequency reduce by background selection. The spectrum depends on the number of individual []

\[\mathbb{E}[Ps_{(x)}] = \sum{x^{*}=x}{x^{*}=1}f_{B}(x)\]

Return:

Array{Float64}: expected rate of neutral alleles frequencies.

Analytical.poissonFixation — Function

poissonFixation(observedValues,λds, λdn)

Divergence sampling from Poisson distribution. The expected neutral and selected fixations are subset through their relative expected rates (Analytical.fixNeut, Analytical.fixNegB, Analytical.fixPosSim). Empirical values are used are used to simulate the locus L along a branch of time T from which the expected Ds and Dn raw count estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily given the success rate $\lambda$ in the distribution.

Arguments

observedValues::Array{Int64,1}: Array containing the total observed divergence.

λds: expected neutral fixations rate.

λdn: expected selected fixations rate.

Returns

Array{Int64,1} containing the expected count of neutral and selected fixations.

Analytical.poissonPolymorphism — Function

poissonPolymorphism(observedValues,λps,λpn)

Polymorphism sampling from Poisson distributions. The total expected neutral and selected polimorphism are subset through the relative expected rates at the frequency spectrum (Analytical.fixNeut, Analytical.DiscSFSNeutDown,). Empirical sfs are used to simulate the locus L along a branch of time T from which the expected Ps and Pn raw count are estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily from the whole sfs given each frequency success rate $\lambda$ in the distribution.

Arguments

observedValues::Array{Int64,1}: Array containing the total observed divergence.

λps: expected neutral site frequency spectrum rate.

λpn: expected selected site frequency spectrum rate.

Returns

Array{Int64,1} containing the expected total count of neutral and selected polymorphism.

Analytical.alphaByFrequencies — Function

alphaByFrequencies(gammaL,gammaH,pposL,pposH,observedData,nopos)

Analytical α(x) estimation. We used the expected rates of divergence and polymorphism to approach the asympotic value accouting for background selection, weakly and strong positive selection. α(x) can be estimated taking into account the role of positive selected alleles or not. In this way we explore the role of linkage to deleterious alleles in the coding region. Solve α(x) from the expectation rates:

\[\mathbb{E}[\alpha_{x}] = 1 - \left(\frac{\mathbb{E}[D_{s}]}{\mathbb{E}[D_{N}]}\frac{\mathbb{E}[P_{N}]}{\mathbb{E}[P_{S}]}\right)\]

Arguments

gammaL::Int64: strength of weakly positive selection

gammaH::Int64: strength of strong positive selection

pposL::Float64: probability of weakly selected allele

pposH::Float64: probability of strong selected allele

observedData::Array{Any,1}: Array containing the total observed divergence, polymorphism and site frequency spectrum.

nopos::String("pos","nopos","both"): string to perform α(x) account or not for both positive selective alleles.

Returns

Tuple{Array{Float64,1},Array{Float64,2}} containing α(x) and the summary statistics array (Ds,Dn,Ps,Pn,α).

Analytical.phiReduction — Function

phiReduction(gamma,ppos)

Reduction in fixation probabilty due to background selection and linkage. The formulas used have been subjected to several theoretical works (Charlesworth B., 1994, Hudson et al., 1995, Nordborg et al. 1995, Barton NH., 1995).

The fixation probabilty of selected alleles are reduce by a factor $\phi$:

\[\phi(t,s) = e^{[\frac{-2\mu}{t(1+\frac{rL}{t}+\frac{2s}{t})}]}\]

Multiplying across all deleterious linkes sites, we find:

\[\Phi = \prod_{1}^{L} = \phi(t,s) = e^{\frac{-2t\mu(\psi[1,\frac{r+2s+L}{r}] - \psi[1,\frac{r(L+1)+2s+t}{r}])}{r^2}}\]

\[\phi(t,s) = e^{\frac{-2t\mu(\psi[1,\frac{r+2s+L}{r}] - \psi[1,\frac{r(L+1)+2s+t}{r}])}{r^2}}\]

Arguments

gamma::Int64: selection coefficient.

Returns

Float64: expected rate of positive fixations under background selection.

Analytical.adap — Constant

adap( 
	gam_neg::Int64,
	gL::Int64,	
	gH::Int64,
	alLow::Float64,
	alTot::Float64,
	theta_f::Float64,
	theta_mid_neutral::Float64,
	al::Float64,
	be::Float64,
	B::Float64,
	bRange::Array{Float64,1}
	pposL::Float64,
	pposH::Float64,
	N::Int64,
	n::Int64,
	Lf::Int64,
	rho::Float64,
	TE::Float64,
)

Mutable structure containing the variables required to solve the analytical approach. All the functions are solve using the internal values of the structure. For this reason, adap is the only exported variable. Adap should be change before the perform the analytical approach, in other case, $\alpha_{(x)}$ will be solve with the default values.

Parameters

gam_neg::Int64:

gL::Int64:

gH::Int64:

alLow::Float64:

alTot::Float64:

theta_f::Float64:

theta_mid_neutral::Float64:

al::Float64:

be::Float64:

B::Float64:

bRange::Array{Float64,1}:

pposL::Float64:

pposH::Float64:

N::Int64:

n::Int64:

Lf::Int64:

rho::Float64:

TE::Float64:

Installation

To install our module we highly recommend to use LTS official Julia binaries. If is your first time using Julia, you can easily export the Julia bin through export PATH="/path/to/directory/julia-1.v.v/bin:$PATH" in your shell. Since we use scipy to solve equations, the package depends on PyCall:

julia -e 'using Pkg;Pkg.add("PyCall");using PyCall;pyimport_conda("scipy.optimize", "scipy")'
julia -e 'using Pkg;Pkg.add(PackageSpec(path="https://github.com/jmurga/Analytical.jl"))'

Or from Pkg REPL (by pressing ] at Julia interpreter):

add PyCall
add "https://github.com/jmurga/Analytical.jl"
# exits from Pkg REPL and run
using PyCall
pyimport_conda("scipy.optimize", "scipy")

In addition we provide a Docker image based on Debian including Julia and Jupyter notebook. You can access to Debian system or just to Jupyter pulling the image from dockerhub. Remember to link the folder /analysis with any folder at your home to save the results:

# Pull the image
docker pull jmurga/mktest
# Run docker bash interactive session linking to some local volume to export data. 
docker run -i -t -v ${HOME}/folderPath:/analysis/folder  jmurga/mktest
# Run only jupyter notebook from docker image. Change the port if 8888 is already used
docker run -i -t -v ${HOME}/folderPath:/analysis/folder -p 8888:8888 jmurga/mktest /bin/bash -c "jupyter-lab --ip='*' --port=8888 --no-browser --allow-root"

Dependencies

References