Analytical approach to $\alpha_{x}$

Analytical estimation of $\alpha_{x}$ asympotic accounting for linkage. We explore the impact of linkage and background selection at positive selected alleles sites.

Analytical.fixNeut — Function

fixNeut()

Expected neutral fixations rate reduce by a background selection value. It takes into accoun the ammount probability of being synonymous.

math \mathbb{E}\left[D_{s}\rigth] = \left(1 - p_{-} - p_{+}) \cdot B \cdot \frac{1}{2N}\rigth)`

Returns

Expected rate of neutral fixations: Float64

Analytical.fixNegB — Function

fixNegB(ppos)

Expected fixation rate from negative DFE

\[\mathbb{E}\left[D_{n-}\right] = p_{-} \left(2^-\alpha \cdot \beta^\alpha \cdot \left(-\zeta\left[\alpha,2+\beta/2] + \zeta\left(\left[\alpha,1/2*\left(\left(2-\fract{1}{N+\beta}\rigth)\rigth]\rigth)\rigth)\]

Arguments

- ```ppos::Float64```: selection coefficient

Negative fixations.

Returns

Expected rate of fixations from negative DFE in a Float64

Analytical.pFix — Function

fixNegB(pFix)

Expected positive fixation rate

\[\mathbb{E}\left[D_{s}\rigth] = \left(1 - p_{-} - p_{+}) \cdot B \cdot \frac{1}{2N}\rigth)\]

Arguments

- ```ppos::Float64```: selection coefficient

Returns

Rate of neutral fixation in a Float64