adap variable

adap is the only variable exported from Analytical module. It is a Mutable structure contaning the variables required to solve the 4 analytical approach. Any value could be easily change reassing. adap should be change before the execution, in other case, $\alpha_{(x)$ will be solve with the default values. To change all the values at once, you can use Analytical.changeParameters in order to set a model.

Analytical.adap — Constant

adap( 
	gam_neg::Int64,
	gL::Int64,	
	gH::Int64,
	alLow::Float64,
	alTot::Float64,
	theta_f::Float64,
	theta_mid_neutral::Float64,
	al::Float64,
	be::Float64,
	B::Float64,
	bRange::Array{Float64,1}
	pposL::Float64,
	pposH::Float64,
	N::Int64,
	n::Int64,
	Lf::Int64,
	rho::Float64,
	TE::Float64,
)

Mutable structure containing the variables required to solve the analytical approach. All the functions are solve using the internal values of the structure. For this reason, adap is the only exported variable. Adap should be change before the perform the analytical approach, in other case, $\alpha_{(x)}$ will be solve with the default values.

Parameters

gam_neg::Int64:

gL::Int64:

gH::Int64:

alLow::Float64:

alTot::Float64:

theta_f::Float64:

theta_mid_neutral::Float64:

al::Float64:

be::Float64:

B::Float64:

bRange::Array{Float64,1}:

pposL::Float64:

pposH::Float64:

N::Int64:

n::Int64:

Lf::Int64:

rho::Float64:

TE::Float64:

Analytical estimation

Fixations

Analytical.fixNeut — Function

fixNeut()

Expected neutral fixations rate reduce by a background selection value.

\[\mathbb{E}[D_{s}] = (1 - p_{-} - p_{+}) B \frac{1}{2N}\]

Returns

Float64: expected rate of neutral fixations.

Analytical.fixNegB — Function

fixNegB(ppos)

Expected fixation rate from negative DFE.

\[\mathbb{E}[D_{n-}] = p_{-}\left(2^-\alpha\beta^\alpha\left(-\zeta[\alpha,\frac{2+\beta}{2}] + \zeta[\alpha,1/2(2-\frac{1}{N+\beta})]\right)\right)\]

Arguments

ppos::Float64: selection coefficient.

Returns

Float64: expected rate of fixations from negative DFE.

Analytical.pFix — Function

pFix()

Expected positive fixation rate.

\[\mathbb{E}[D_{n+}] = p_{+} \cdot B \cdot (1 - e^{(-2s)})\]

Arguments

ppos::Float64: selection coefficient.

Returns

Float64: expected rate of positive fixation.

Analytical.fixPosSim — Function

fixPosSim(gamma,ppos)

Expected positive fixations rate reduced due to the impact of background selection and linkage. The probabilty of fixation of positively selected alleles is reduced by a factor Φ across all deleterious linked sites Analytical.phiReduction.

\[\mathbb{E}[D_{n+}] = \Phi \cdot \mathbb{E}[D_{n+}]\]

Arguments

ppos::Float64: selection coefficient

Returns

Float64: expected rate of positive fixations under background selection

Polymorphism

Analytical.DiscSFSNeutDown — Function

DiscSFSNeutDown()

Expected rate of neutral allele frequency reduce by background selection. The spectrum depends on the number of individual []

\[\mathbb{E}[Ps_{(x)}] = \sum{x^{*}=x}{x^{*}=1}f_{B}(x)\]

Return:

Array{Float64}: expected rate of neutral alleles frequencies.

Summary statistics

Analytical.poissonFixation — Function

poissonFixation(observedValues,λds, λdn)

Divergence sampling from Poisson distribution. The expected neutral and selected fixations are subset through their relative expected rates (Analytical.fixNeut, Analytical.fixNegB, Analytical.fixPosSim). Empirical values are used are used to simulate the locus L along a branch of time T from which the expected Ds and Dn raw count estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily given the success rate $\lambda$ in the distribution.

Arguments

observedValues::Array{Int64,1}: Array containing the total observed divergence.

λds: expected neutral fixations rate.

λdn: expected selected fixations rate.

Returns

Array{Int64,1} containing the expected count of neutral and selected fixations.

Analytical.poissonPolymorphism — Function

poissonPolymorphism(observedValues,λps,λpn)

Polymorphism sampling from Poisson distributions. The total expected neutral and selected polimorphism are subset through the relative expected rates at the frequency spectrum (Analytical.fixNeut, Analytical.DiscSFSNeutDown,). Empirical sfs are used to simulate the locus L along a branch of time T from which the expected Ps and Pn raw count are estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily from the whole sfs given each frequency success rate $\lambda$ in the distribution.

Arguments

observedValues::Array{Int64,1}: Array containing the total observed divergence.

λps: expected neutral site frequency spectrum rate.

λpn: expected selected site frequency spectrum rate.

Returns

Array{Int64,1} containing the expected total count of neutral and selected polymorphism.

Analytical.alphaByFrequencies — Function

alphaByFrequencies(gammaL,gammaH,pposL,pposH,observedData,nopos)

Analytical α(x) estimation. We used the expected rates of divergence and polymorphism to approach the asympotic value accouting for background selection, weakly and strong positive selection. α(x) can be estimated taking into account the role of positive selected alleles or not. We solve α(x) from empirical observed values. The values will be use to sample from a Poisson distribution the total counts of polymorphism and divergence using the rates. The mutation rate, the locus length and the time of the branch should be proportional to the observed values.

\[\mathbb{E}[\alpha_{x}] = 1 - \left(\frac{\mathbb{E}[D_{s}]}{\mathbb{E}[D_{N}]}\frac{\mathbb{E}[P_{N}]}{\mathbb{E}[P_{S}]}\right)\]

Arguments

gammaL::Int64: strength of weakly positive selection

gammaH::Int64: strength of strong positive selection

pposL::Float64: probability of weakly selected allele

pposH::Float64: probability of strong selected allele

observedData::Array{Any,1}: Array containing the total observed divergence, polymorphism and site frequency spectrum.

nopos::String("pos","nopos","both"): string to perform α(x) account or not for both positive selective alleles.

Returns

Tuple{Array{Float64,1},Array{Float64,2}} containing α(x) and the summary statistics array (Ds,Dn,Ps,Pn,α).