Model parameters

adap is the only variable exported from Analytical module. It is a Mutable structure contaning the variables required to solve the analytical approach. Any value can be easly changed. Remember adap should be change before the execution, in other case, $\alpha_{(x)}$ will be solve with the default values. To change all the values at once, you can use Analytical.parameters in order to set specific models in a new variable.

Analytical.parameters — Type

Mutable structure containing the variables required to solve the analytical approach. All the functions are solve using the internal values of the structure. You should declare a mutable structure to the perform the analytical estimations.

Parameters

gamNeg::Int64: Selection coefficient for deleterious alleles

gL::Int64: Selection coefficient for weakly benefical alleles

gH::Int64: Selection coefficient for strongly benefical alleles

alLow::Float64: Proportion of α due to weak selection

alTot::Float64: α

thetaF::Float64: Mutation rate defining BGS strength

thetaMidNeutral::Float64: Mutation rate on coding region

al::Float64: DFE shape parameter

be::Float64: DFE scale parameter

B::Float64: BGS strength

bRange::Array{Float64,1}: BGS values to simulate

pposL::Float64: Fixation probabily of weakly beneficial alleles

pposH::Float64: Fixation probabily of strongly beneficial alleles

N::Int64: Population size

n::Int64: Sample size

Lf::Int64: Flanking region length

rho::Float64: Recombination rate

TE::Float64

Analytical.binomialDict — Type

Mutable structure containing the downsampled SFS.

Returns

bn::Dict: SparseMatrixCSC containing the binomial convolution

Analytical.Br — Function

Br(Lmax,theta)

Expected reduction in nucleotide diversity. Explored at Charlesworth B., 1994:

\[\frac{\pi}{\pi_{0}} = e^{\frac{4\muL}{2rL+t}}\]

Arguments

param::parameters

theta::Float64

Returns

Float64: expected reduction in diversity given a non-coding length, mutation rate and defined recombination.

Analytical.setThetaF! — Function

setThetaF!(param)

Find the optimum mutation given the expected reduction in nucleotide diversity (B value) in a locus.

Returns

adap.thetaF::Float64: changes adap.thetaF value.

Analytical.setPpos! — Function

setPpos!(param)

Find the probabilty of positive selected alleles given the model. It solves a equation system taking into account fixations probabilities of weakly and strong beneficial alleles.

Returns

Tuple{Float64,Float64}: weakly and strong beneficial alleles probabilites.

Analytical.binomOp! — Function

binomOp(param)

Binomial convolution to sample the allele frequencies probabilites depending on background selection values and sample size.

Arguments

param::parameters

convolutedSamples::binomialDict

Returns

Array{Float64,2}: convoluted SFS for each B value defined in the model (param.bRange). The estimations are saved at convolutedBn.bn.

Analytical.phiReduction — Function

phiReduction(param,gamma,ppos)

Reduction in fixation probabilty due to background selection and linkage. The formulas used have been subjected to several theoretical works (Charlesworth B., 1994, Hudson et al., 1995, Nordborg et al. 1995, Barton NH., 1995).

The fixation probabilty of selected alleles are reduce by a factor $\phi$:

\[\phi(t,s) = e^{[\frac{-2\mu}{t(1+\frac{rL}{t}+\frac{2s}{t})}]}\]

Multiplying across all deleterious linkes sites, we find:

\[\Phi = \prod_{1}^{L} = \phi(t,s) = e^{\frac{-2t\mu(\psi[1,\frac{r+2s+L}{r}] - \psi[1,\frac{r(L+1)+2s+t}{r}])}{r^2}}\]

\[\phi(t,s) = e^{\frac{-2t\mu(\psi[1,\frac{r+2s+L}{r}] - \psi[1,\frac{r(L+1)+2s+t}{r}])}{r^2}}\]

Arguments

gamma::Int64: selection coefficient.

Returns

Float64: expected rate of positive fixations under background selection.

Analytical.analyticalAlpha — Function

analyticalAlpha(param, convolutedSamples)

Analytical α(x) estimation. Solve α(x) generally. We used the expected rates of divergence and polymorphism to approach the asympotic value accouting for background selection, weakly and strong positive selection. α(x) can be estimated taking into account the role of positive selected alleles or not. In this way we explore the role of linkage to deleterious alleles in the coding region.

\[\mathbb{E}[\alpha_{x}] = 1 - \left(\frac{\mathbb{E}[D_{s}]}{\mathbb{E}[D_{N}]}\frac{\mathbb{E}[P_{N}]}{\mathbb{E}[P_{S}]}\right)\]

Arguments

param::parameters

convolutedSamples::binomialDict

Returns

Array{Float64,1} α(x).

Analytical estimation

Fixations

Analytical.fixNeut — Function

fixNeut()

Expected neutral fixations rate reduce by B value.

\[\mathbb{E}[D_{s}] = (1 - p_{-} - p_{+}) B \frac{1}{2N}\]

Returns

Float64: expected rate of neutral fixations.

Analytical.fixNegB — Function

fixNegB(ppos)

Expected fixation rate from negative DFE.

\[\mathbb{E}[D_{n-}] = p_{-}\left(2^-\alpha\beta^\alpha\left(-\zeta[\alpha,\frac{2+\beta}{2}] + \zeta[\alpha,1/2(2-\frac{1}{N+\beta})]\right)\right)\]

Arguments

ppos::Float64: positive selected alleles probabilty.

Returns

Float64: expected rate of fixations from negative DFE.

Analytical.pFix — Function

pFix()

Expected positive fixation rate.

\[\mathbb{E}[D_{n+}] = p_{+} \cdot B \cdot (1 - e^{(-2s)})\]

Arguments

ppos::Float64: positive selected alleles probabilty.

Returns

Float64: expected rate of positive fixation.

Analytical.fixPosSim — Function

fixPosSim(gamma,ppos)

Expected positive fixations rate reduced due to the impact of background selection and linkage. The probabilty of fixation of positively selected alleles is reduced by a factor Φ across all deleterious linked sites Analytical.phiReduction.

\[\mathbb{E}[D_{n+}] = \Phi \cdot \mathbb{E}[D_{n+}]\]

Arguments

ppos::Float64: positive selected alleles probabilty.

Returns

Float64: expected rate of positive fixations under background selection

Polymorphism

Analytical.DiscSFSNeutDown — Function

DiscSFSNeutDown()

Expected rate of neutral allele frequency reduce by backgrou nd selection. The spectrum depends on the number of individual []

\[\mathbb{E}[Ps_{(x)}] = \sum{x^{*}=x}{x^{*}=1}f_{B}(x)\]

Return:

Array{Float64}: expected rate of neutral alleles frequencies.

Analytical.DiscSFSSelPosDown — Function

DiscSFSSelPosDown(gammaValue,ppos)

Expected rate of positive selected allele frequency reduce by background selection. The spectrum depends on the number of individuals.

Arguments

gammaValue::Int64: selection strength.

ppos::Float64: positive selected alleles probabilty.

Return:

Array{Float64}: expected positive selected alleles frequencies.

Analytical.DiscSFSSelNegDown — Function

DiscSFSSelNegDown(param,ppos)

Expected rate of positive selected allele frequency reduce by background selection. Spectrum drawn on a gamma DFE. It depends on the number of individuals.

Arguments

ppos::Float64: positive selected alleles probabilty.

Return:

Array{Float64}: expected negative selected alleles frequencies.

Analytical.cumulativeSfs — Function

cumulativeSfs(sfsTemp)

Changing SFS considering all values above a frequency x. The original asymptotic-MK approach takes Pn(x) and Ps(x) as the number of polymorphic sites at frequency x rather than above x, but this approach scales poorly as sample size increases. We define the polymorphic spectrum as stated above since these quantities trivially have the same asymptote but are less affected by changing sample size.

Summary statistics

Analytical.poissonFixation — Function

poissonFixation(observedValues,λds, λdn)

Divergence sampling from Poisson distribution. The expected neutral and selected fixations are subset through their relative expected rates (fixNeut, fixNegB, fixPosSim). Empirical values are used are used to simulate the locus L along a branch of time T from which the expected Ds and Dn raw count estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily given the success rate $\lambda$ in the distribution.

\[\mathbb{E}[D_N] = X \in Poisson\left(\lambda = D \times \left[\frac{\mathbb{E}[D_+] + \mathbb{E}[D_-]}{\mathbb{E}[D_+] + \mathbb{E}[D_-] + \mathbb{E}[D_0]}\right]\right)\]

\[\mathbb{E}[D_S] = X \in Poisson\left(\lambda = D \times \left[\frac{\mathbb{E}[D_0]}{\mathbb{E}[D_+] + \mathbb{E}[D_-] + \mathbb{E}[D_0]}\right]\right)\]

Arguments

observedValues::Array: Array containing the total observed divergence.

λds::Float64: expected neutral fixations rate.

λdn::Float64: expected selected fixations rate.

Returns

Array{Int64,1} containing the expected count of neutral and selected fixations.

Analytical.poissonPolymorphism — Function

poissonPolymorphism(observedValues,λps,λpn)

Polymorphism sampling from Poisson distributions. The total expected neutral and selected polimorphism are subset through the relative expected rates at the frequency spectrum (fixNeut, DiscSFSNeutDown,). Empirical sfs are used to simulate the locus L along a branch of time T from which the expected Ps and Pn raw count are estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily from the whole sfs given each frequency success rate $\lambda$ in the distribution.

The success rate managing the Poisson distribution by the observed count each frequency. We considered both sampling variance and process variance is affecting the number of variable alleles we sample from SFS. This variance arises from the random mutation-fixation process along the branch. To incorporate this variance we do one sample per frequency-bin and use the total sampled variation and the SFS for the summary statistics.

\[\mathbb{E}[P_N] = \sum_{x=0}^{x=1} X \in Poisson\left(\lambda = SFS_{(x)} \times \left[\frac{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}]}{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}] + \mathbb{E}[P_{0(x)}]}\right]\right)\]

\[\mathbb{E}[P_S] = \sum_{x=0}^{x=1} X \in Poisson\left(\lambda = SFS_{(x)} \times \left[\frac{\mathbb{E}[P_{0(x)}]}{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}] + \mathbb{E}[P_{0(x)}]}\right]\right)\]

Arguments

observedValues::Array{Int64,1}: Array containing the total observed divergence.

λps::Array{Float64,1}: expected neutral site frequency spectrum rate.

λpn::Array{Float64,1}: expected selected site frequency spectrum rate.

Returns

Array{Int64,2} containing the expected total count of neutral and selected polymorphism.

Analytical.sampledAlpha — Function

sampledAlpha(observedValues,λds, λdn)

Ouput the expected values from the Poisson sampling process. Please check poissonFixation and poissonPolymorphism to understand the samplingn process. α(x) is estimated through the expected values of Dn, Ds, Pn and Ps.

Arguments

param::parameters: Array containing the total observed divergence.

d::Array: observed divergence.

afs::Array: observed polymorphism.

λdiv::Array{Float64,2}: expected fixations rate.

λdiv::Array{Float64,2}: expected site frequency spectrum rates.

Returns

αS,expDn,expDs,expPn,expPs,ssAlpha

Array{Int64,2} containing α(x) values.

Array{Int64,1} expected non-synonymous divergence.

Array{Int64,1} expected synonymous divergence.

Array{Int64,1} expected non-synonymous polymorphism.

Array{Int64,1} expected synonymous polymorphism.

Array{Int64,1} expected synonymous polymorphism.

Array{Int64,1} expected synonymous polymorphism.

Array{Int64,2} containing α(x) binned values.

Analytical.iterRates — Function

iterRates(param::parameters,afac::Float64,bfac::Float64,alTot::Float64,alLow::Float64,divergence::Array,sfs::Array)

Estimating rates given a model for all B range.

Arguments

param::parameters

convolutedSamples::binomialDict

alTot::Float64

alLow::Float64

gH::Int64

gL::Int64

gamNeg::Int64

afac::Float64

ρ::Float64

θ::Float64

Output

Array{Float64,2}

Analytical.gettingRates — Function

gettingRates(gammaL,gammaH,pposL,pposH,observedData,nopos)

Estimating analytical rates of fixation and polymorphism to approach α value accouting for background selection, weakly and strong positive selection. Output values will be used to sample from a Poisson distribution the total counts of polymorphism and divergence using observed data.

Arguments

param::parameters

cnvBinom::SparseMatrixCSC{Float64,Int64}

Returns

Array{Float64,2} containing solved model, fixation and polymorphic rates

Analytical.summaryStatsFromRates — Function

summaryStatsFromRates(param::parameters,rates::JLD2.JLDFile,analysisFolder::String,summstatSize::Int64,replicas::Int64,bootstrap::Bool)

Estimate summary statistics using observed data and analytical rates. analysisFolder will check for the SFS and divergence file and will be used to output summary statistics

Arguments

param::parameters : Mutable structure containing the models

rates::JLD2.JLDFile : HDF5 containing solved models, fixation and polymorphic rates

analysisFolder::String : Folder containing the SFS and divergence files. It will be used to output the observed data and summary estatistics.

summstatSize::Int64 : Number of summary statistics

replicas::Int64 : Number of bootstrap replicas

bootstrap::Bool : Boolean to perform or not bootstrapping

Output

Obserded data and summary statistics to ABC inference

Inference tools

Analytical.parseSfs — Function

parseSfs(;data,output,sfsColumns,divColumns)

Function to parse polymorphism and divergence by subset of genes. The input data is based on supplementary material described at Uricchio et al. 2019. Please be sure the file is tabulated.

GeneId	Pn	DAF seppareted by commas	Ps	DAF separated by commas	Dn	Ds

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

Arguments

data: String or Array of strings containing files names with full path.

output::String: path to save file. Containing one file per input file.

sfsColumns::Array{Int64,1}: non-synonymous and synonymous daf columns. Please introduce first the non-synonymous number.

divColumns::Array{Int64,1}: non-synonymous and synonymous divergence columns. Please introduce first the non-synonymous number.

Returns

Array{Float64,1}: α values

Array{Float64,2}: Site Frequency Spectrum

Array{Float64,1}: Synonymous and non-synonymous divergence counts

Analytical.ABCreg — Function

ABCreg(analysisFolder, replicas, S, tol, abcreg)

Performing ABC inference using ABCreg. Please, be sure your analysisFolder contain the files alphas.txt and summaries.txt produced by Analytical.summaryStatsFromRates()

Arguments

analysisFolder::String : Folder containing the observed data and summary estatistics. It will be used to output the posterior distributions

S::Int64 : Number of summary stastitics to perform the inference.

tol::Float64 : Tolerance value. It define the number of accepted value at ABC inference

abcreg::String : Path to ABCreg binary

Output

Files containing posterior distributions from ABCreg