Model parameters

adap is the only variable exported from Analytical module. It is a Mutable structure contaning the variables required to solve the analytical approach. Any value can be easly changed. Remember adap should be change before the execution, in other case, $\alpha_{(x)}$ will be solve with the default values. To change all the values at once, you can use Analytical.parameters in order to set specific models in a new variable.

Analytical.parameters — Type

parameters(
	gam_neg::Int64,
	gL::Int64,
	gH::Int64,
	alLow::Float64,
	alTot::Float64,
	theta_f::Float64,
	theta_mid_neutral::Float64,
	al::Float64,
	be::Float64,
	B::Float64,
	bRange::Array{Float64,1}
	pposL::Float64,
	pposH::Float64,
	N::Int64,
	n::Int64,
	Lf::Int64,
	rho::Float64,
	TE::Float64,
)

Mutable structure containing the variables required to solve the analytical approach. All the functions are solve using the internal values of the structure. For this reason, adap is the only exported variable. Adap should be change before the perform the analytical approach, in other case, $\alpha_{(x)}$ will be solve with the default values.

Parameters

gam_neg::Int64:

gL::Int64:

gH::Int64:

alLow::Float64:

alTot::Float64:

theta_f::Float64:

theta_mid_neutral::Float64:

al::Float64:

be::Float64:

B::Float64:

bRange::Array{Float64,1}:

pposL::Float64:

pposH::Float64:

N::Int64:

n::Int64:

Lf::Int64:

rho::Float64:

TE::Float64:

Analytical.Br — Function

Br(Lmax,theta)

Expected reduction in nucleotide diversity. Explored at Charlesworth B., 1994:

\[\frac{\pi}{\pi_{0}} = e^{\frac{4\muL}{2rL+t}}\]

Arguments

Lmax::Int64: non-coding flaking length

theta::Float64

Returns

Float64: expected reduction in diversity given a non-coding length, mutation rate and defined recombination.

Analytical.set_theta_f! — Function

set_theta_f!(param)

Find the optimum mutation given the expected reduction in nucleotide diversity (B value) in a locus.

Returns

adap.theta_f::Float64: changes adap.theta_f value.

Missing docstring.

Missing docstring for Analytical.setPpos!. Check Documenter's build log for details.

Analytical.binomOp! — Function

binomOp(param)

Site Frequency Spectrum convolution depeding on background selection values. Pass the SFS to a binomial distribution to sample the allele frequencies probabilites.

Returns

Array{Float64,2}: convoluted SFS given a B value. It will be saved at adap.bn.

Analytical estimation

Fixations

Analytical.fixNeut — Function

fixNeut()

Expected neutral fixations rate reduce by a background selection value.

\[\mathbb{E}[D_{s}] = (1 - p_{-} - p_{+}) B \frac{1}{2N}\]

Returns

Float64: expected rate of neutral fixations.

Analytical.fixNegB — Function

fixNegB(ppos)

Expected fixation rate from negative DFE.

\[\mathbb{E}[D_{n-}] = p_{-}\left(2^-\alpha\beta^\alpha\left(-\zeta[\alpha,\frac{2+\beta}{2}] + \zeta[\alpha,1/2(2-\frac{1}{N+\beta})]\right)\right)\]

Arguments

ppos::Float64: positive selected alleles probabilty.

Returns

Float64: expected rate of fixations from negative DFE.

Analytical.pFix — Function

pFix()

Expected positive fixation rate.

\[\mathbb{E}[D_{n+}] = p_{+} \cdot B \cdot (1 - e^{(-2s)})\]

Arguments

ppos::Float64: positive selected alleles probabilty.

Returns

Float64: expected rate of positive fixation.

Analytical.fixPosSim — Function

fixPosSim(gamma,ppos)

Expected positive fixations rate reduced due to the impact of background selection and linkage. The probabilty of fixation of positively selected alleles is reduced by a factor Φ across all deleterious linked sites Analytical.phiReduction.

\[\mathbb{E}[D_{n+}] = \Phi \cdot \mathbb{E}[D_{n+}]\]

Arguments

ppos::Float64: positive selected alleles probabilty.

Returns

Float64: expected rate of positive fixations under background selection

Polymorphism

Analytical.DiscSFSNeutDown — Function

DiscSFSNeutDown()

Expected rate of neutral allele frequency reduce by background selection. The spectrum depends on the number of individual []

\[\mathbb{E}[Ps_{(x)}] = \sum{x^{*}=x}{x^{*}=1}f_{B}(x)\]

Return:

Array{Float64}: expected rate of neutral alleles frequencies.

Analytical.DiscSFSSelPosDown — Function

DiscSFSSelPosDown(gammaValue,ppos)

Expected rate of positive selected allele frequency reduce by background selection. The spectrum depends on the number of individuals.

Arguments

gammaValue::Int64: selection strength.

ppos::Float64: positive selected alleles probabilty.

Return:

Array{Float64}: expected positive selected alleles frequencies.

Analytical.DiscSFSSelNegDown — Function

DiscSFSSelNegDown(param,ppos)

Expected rate of positive selected allele frequency reduce by background selection. Spectrum drawn on a gamma DFE. It depends on the number of individuals.

Arguments

ppos::Float64: positive selected alleles probabilty.

Return:

Array{Float64}: expected negative selected alleles frequencies.

Analytical.cumulativeSfs — Function

cumulativeSfs(sfsTemp)

Changing SFS considering all values above a frequency x. The original asymptotic-MK approach takes Pn(x) and Ps(x) as the number of polymorphic sites at frequency x rather than above x, but this approach scales poorly as sample size increases. We define the polymorphic spectrum as stated above since these quantities trivially have the same asymptote but are less affected by changing sample size.

Summary statistics

Analytical.poissonFixation — Function

poissonFixation(observedValues,λds, λdn)

Divergence sampling from Poisson distribution. The expected neutral and selected fixations are subset through their relative expected rates (Analytical.fixNeut, Analytical.fixNegB, Analytical.fixPosSim). Empirical values are used are used to simulate the locus L along a branch of time T from which the expected Ds and Dn raw count estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily given the success rate $\lambda$ in the distribution.

\[\mathbb{E}[D_N] = X \in Poisson\left(\lambda = D \times \left[\frac{\mathbb{E}[D_+] + \mathbb{E}[D_-]}{\mathbb{E}[D_+] + \mathbb{E}[D_-] + \mathbb{E}[D_0]}\right]\right)\]

\[\mathbb{E}[D_S] = X \in Poisson\left(\lambda = D \times \left[\frac{\mathbb{E}[D_0]}{\mathbb{E}[D_+] + \mathbb{E}[D_-] + \mathbb{E}[D_0]}\right]\right)\]

Arguments

observedValues::Array{Int64,1}: Array containing the total observed divergence.

λds: expected neutral fixations rate.

λdn: expected selected fixations rate.

Returns

Array{Int64,1} containing the expected count of neutral and selected fixations.

Analytical.poissonPolymorphism — Function

poissonPolymorphism(observedValues,λps,λpn)

Polymorphism sampling from Poisson distributions. The total expected neutral and selected polimorphism are subset through the relative expected rates at the frequency spectrum (Analytical.fixNeut, Analytical.DiscSFSNeutDown,). Empirical sfs are used to simulate the locus L along a branch of time T from which the expected Ps and Pn raw count are estimated given the mutation rate ($\mu$). Random number generation is used to subset samples arbitrarily from the whole sfs given each frequency success rate $\lambda$ in the distribution.

The success rate managing the Poisson distribution by the observed count each frequency. We considered both sampling variance and process variance is affecting the number of variable alleles we sample from SFS. This variance arises from the random mutation-fixation process along the branch. To incorporate this variance we do one sample per frequency-bin and use the total sampled variation and the SFS for the summary statistics.

\[\mathbb{E}[P_N] = \sum_{x=0}^{x=1} X \in Poisson\left(\lambda = SFS_{(x)} \times \left[\frac{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}]}{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}] + \mathbb{E}[P_{0(x)}]}\right]\right)\]

\[\mathbb{E}[P_S] = \sum_{x=0}^{x=1} X \in Poisson\left(\lambda = SFS_{(x)} \times \left[\frac{\mathbb{E}[P_{0(x)}]}{\mathbb{E}[P_{+(x)}] + \mathbb{E}[P_{-(x)}] + \mathbb{E}[P_{0(x)}]}\right]\right)\]

Arguments

observedValues::Array{Int64,1}: Array containing the total observed divergence.

λps: expected neutral site frequency spectrum rate.

λpn: expected selected site frequency spectrum rate.

Returns

Array{Int64,1} containing the expected total count of neutral and selected polymorphism.

Analytical.alphaByFrequencies — Function

alphaByFrequencies(gammaL,gammaH,pposL,pposH,observedData,nopos)

Analytical α(x) estimation. We used the expected rates of divergence and polymorphism to approach the asympotic value accouting for background selection, weakly and strong positive selection. α(x) can be estimated taking into account the role of positive selected alleles or not. We solve α(x) from empirical observed values. The values will be use to sample from a Poisson distribution the total counts of polymorphism and divergence using the rates. The mutation rate, the locus length and the time of the branch should be proportional to the observed values.

\[\mathbb{E}[\alpha_{x}] = 1 - \left(\frac{\mathbb{E}[D_{s}]}{\mathbb{E}[D_{N}]}\frac{\mathbb{E}[P_{N}]}{\mathbb{E}[P_{S}]}\right)\]

Arguments

gammaL::Int64: strength of weakly positive selection

gammaH::Int64: strength of strong positive selection

pposL::Float64: probability of weakly selected allele

pposH::Float64: probability of strong selected allele

observedData::Array{Any,1}: Array containing the total observed divergence, polymorphism and site frequency spectrum.

nopos::String("pos","nopos","both"): string to perform α(x) account or not for both positive selective alleles.

Returns

Tuple{Array{Float64,1},Array{Float64,2}} containing α(x) and the summary statistics array (Ds,Dn,Ps,Pn,α).

Inference tools

Analytical.parseSfs — Function

parseSfs(;data,output,sfsColumns,divColumns)

Function to parse polymorphism and divergence by subset of genes. The input data is based on supplementary material described at Uricchio et al. 2019.

GeneId	Pn	DAF seppareted by commas	Ps	DAF separated by commas	Dn	Ds

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

XXXX	0	,0,0,0,0,0,0,0,0	0	,0,0,0,0,0,0,0,0	0	0

Arguments

data: String or Array of strings containing files names with full path.

output::String: path to save file. Containing one file per input file.

sfsColumns::Array{Int64,1}: non-synonymous and synonymous daf columns. Please introduce first the non-synonymous number.

divColumns::Array{Int64,1}: non-synonymous and synonymous divergence columns. Please introduce first the non-synonymous number.

Returns

Array{Array{Int64,N} where N,1}: Array of arrays containing the total polymorphic sites (1), total Site Frequency Spectrum (2) and total divergence (3). Each array contains one row/column per file.

File writed in output

Analytical.ABCreg — Function

ABCreg(data, prior, nparams, nsummaries, outputPath, outputPrefix,tolerance, regressionMode,regPath)

Julia to execute ABCreg. You should take into account that any other ABC could be used once the prior distributions are done. If you consider to use ABCreg please cite the publication and compile it in your system.

Arguments

data::String: Observed data. Produced by parseSfs.

output::String: path to save file. ABCreg will produce one file per lines inside data.

nparams::Int64: number of parameters in prior distributions.

nsummaries::Int64: number of observed summary statistics.

outputPath::String: output path.

outputPrefix::String: output prefix name.

tolerance::Float64: tolerance for rejection acceptance in Euclidian distance.

regressionMode::String: transformation (T or L).

Returns

Array{Array{Int64,N} where N,1}: Array of arrays containing the total polymorphic sites (1), total Site Frequency Spectrum (2) and total divergence (3). Each array contains one row/column per file.

File writed in output

Analytical.meanQ — Function

meanQ(x,columns)

Function to retrieve mean and quantiles (95%) from posterior distributions.

Arguments

x::Array{Float64,2}: posterior distribution.

columns::Array{Int64,1}: columns to process.

Returns

Array{Array{Float64,2},1}: Array of array containing mean and quantiles by posterior distribution. Each array contains $\alpha_{S}$, $\alpha_{W}$ and $\alpha$ information by column.